ai_service/pyproject.toml
[project]
name = "ai-itinerary-service"
version = "0.1.0"
description = "AI service for travel itinerary extraction and generation"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "anthropic>=0.42",
    "yt-dlp>=2024.12",
    "instaloader>=4.13",
    "faster-whisper>=1.1",
    "Pillow>=10.4",
    "beautifulsoup4>=4.12",
    "readability-lxml>=0.8",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    "eval_type_backport>=0.2.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
